tests/submission/Infraware/Forms/contents/Forms/Not Use/keygen_disabled.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>keygen_disabled</h3>
	</p>
	
	<li>The keygen control be disabled when the document was loaded.</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
	 <p><keygen name="key" id='keygen_id'></keygen></p>
	</form>

	<script>
		
		var keygen = document.getElementById("keygen_id");
		keygen.disabled = true;

		if (typeof(keygen.disabled) == "boolean") {
			if (keygen.disabled != true) {
				test(function() {
					assert_true(false, "disabled attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("disabled attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
